Lot Sales added

+++ MMBSeaFoods/src/application/Models/Fish_Lot.java
@@ -11,7 +11,7 @@ public class Fish_Lot {
     private double other_fees; 
     private double brokerFee;  
     private double buying_price;
-    private int Sold_Weight;
+    private double Sold_Weight;
     private double Sold_price;
     private int Sold_To;
     private String display_Name;
@@ -75,10 +75,10 @@ public class Fish_Lot {
 	public void setBuying_price(double buying_price) {
 		this.buying_price = buying_price;
 	}
-	public int getSold_Weight() {
+	public double getSold_Weight() {
 		return Sold_Weight;
 	}
-	public void setSold_Weight(int sold_Weight) {
+	public void setSold_Weight(double sold_Weight) {
 		Sold_Weight = sold_Weight;
 	}
 	public double getSold_price() {
